+++ ChangeLog.rst
@@ -2,11 +2,14 @@
 Scene Manager Version History
 =============================
 
-
 1.92
 -----
 * **New:** add/remove user functions added
 * **Update:** IMPORTANT user preferences (smSettings.json) structure changed. Delete old preference data under user/Documents
+* **Update:** method for getting necessary scene paths has re-written
+* **Update:** Documentation and Docstring updates
+* **Update:** Various code clean-ups
+* **Update:** Added .tif extension to the imageViewer
 
 1.91
 ----